# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.5.0",
    "title": "Kubero",
    "description": "Kubero is a web-based tool deploy applications on a Kubernetes clusters. It provides a simple and intuitive interface to manage your clusters, applications, and pipelines."
  },
  "servers": [
    {
      "url": "http://localhost:2000/api"
    }
  ],
  "paths": {
...
    "/cli/pipelines": {
      "post": {
        "tags": [
          "CLI"
        ],
        "summary": "Create a new pipeline",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": {
              "type": "http",
              "scheme": "bearer",
              "bearerFormat": "JWT"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "git": {
                    "example": "any"
                  },
                  "buildpack": {
                    "example": "any"
                  },
                  "pipelineName": {
                    "example": "any"
                  },
                  "domain": {
                    "example": "any"
                  },
                  "phases": {
                    "example": "any"
                  },
                  "reviewapps": {
                    "example": "any"
                  },
                  "dockerimage": {
                    "example": "any"
                  },
                  "deploymentstrategy": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "CLI"
        ],
        "summary": "Get a list of available pipelines",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": {
              "type": "http",
              "scheme": "bearer",
              "bearerFormat": "JWT"
            }
          }
        ]
      }
    },
    "/pipelines": {
      "post": {
        "tags": [
          "Pipeline"
        ],
        "summary": "Create a new pipeline",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "pipelineName": {
                  "example": "any"
                },
                "domain": {
                  "example": "any"
                },
                "phases": {
                  "example": "any"
                },
                "buildpack": {
                  "example": "any"
                },
                "reviewapps": {
                  "example": "any"
                },
                "git": {
                  "example": "any"
                },
                "dockerimage": {
                  "example": "any"
                },
                "deploymentstrategy": {
                  "example": "any"
                }
              }
            },
            "description": "Pipeline object",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "pipelineName": {
                    "example": "any"
                  },
                  "domain": {
                    "example": "any"
                  },
                  "phases": {
                    "example": "any"
                  },
                  "buildpack": {
                    "example": "any"
                  },
                  "reviewapps": {
                    "example": "any"
                  },
                  "git": {
                    "example": "any"
                  },
                  "dockerimage": {
                    "example": "any"
                  },
                  "deploymentstrategy": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Pipeline"
        ],
        "summary": "Get a list of available pipelines",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/cli/pipelines/{pipeline}": {
      "get": {
        "tags": [
          "CLI"
        ],
        "summary": "Get a pipeline",
        "description": "",
        "parameters": [
          {
            "name": "pipeline",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": {
              "type": "http",
              "scheme": "bearer",
              "bearerFormat": "JWT"
            }
          }
        ]
      },
      "delete": {
        "tags": [
          "CLI"
        ],
        "summary": "Delete a pipeline",
        "description": "",
        "parameters": [
          {
            "name": "pipeline",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": {
              "type": "http",
              "scheme": "bearer",
              "bearerFormat": "JWT"
            }
          }
        ]
      }
    },
    "/pipelines/{pipeline}": {
      "get": {
        "tags": [
          "Pipeline"
        ],
        "summary": "Get a pipeline",
        "description": "",
        "parameters": [
          {
            "name": "pipeline",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Pipeline"
        ],
        "summary": "Delete a pipeline",
        "description": "",
        "parameters": [
          {
            "name": "pipeline",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/pipelines/{pipeline}/{phase}/{app}": {
      "delete": {
        "tags": [
          "Pipeline"
        ],
        "summary": "Delete an app",
        "description": "",
        "parameters": [
          {
            "name": "pipeline",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "phase",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "app",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Pipeline"
        ],
        "summary": "Update an app in a specific pipeline",
        "description": "",
        "parameters": [
          {
            "name": "pipeline",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "phase",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "app",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "buildpack": {
                  "example": "any"
                },
                "deploymentstrategy": {
                  "example": "any"
                },
                "gitrepo": {
                  "example": "any"
                },
                "branch": {
                  "example": "any"
                },
                "autodeploy": {
                  "example": "any"
                },
                "domain": {
                  "example": "any"
                },
                "podsize": {
                  "example": "any"
                },
                "autoscale": {
                  "example": "any"
                },
                "envvars": {
                  "example": "any"
                },
                "image": {
                  "example": "any"
                },
                "web": {
                  "example": "any"
                },
                "worker": {
                  "example": "any"
                },
                "cronjobs": {
                  "example": "any"
                },
                "addons": {
                  "example": "any"
                },
                "resourceVersion": {
                  "example": "any"
                }
              }
            },
            "description": "App object",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "buildpack": {
                    "example": "any"
                  },
                  "deploymentstrategy": {
                    "example": "any"
                  },
                  "gitrepo": {
                    "example": "any"
                  },
                  "branch": {
                    "example": "any"
                  },
                  "autodeploy": {
                    "example": "any"
                  },
                  "domain": {
                    "example": "any"
                  },
                  "podsize": {
                    "example": "any"
                  },
                  "autoscale": {
                    "example": "any"
                  },
                  "envvars": {
                    "example": "any"
                  },
                  "image": {
                    "example": "any"
                  },
                  "web": {
                    "example": "any"
                  },
                  "worker": {
                    "example": "any"
                  },
                  "cronjobs": {
                    "example": "any"
                  },
                  "addons": {
                    "example": "any"
                  },
                  "resourceVersion": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Pipeline"
        ],
        "summary": "Get app details",
        "description": "",
        "parameters": [
          {
            "name": "pipeline",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "phase",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "app",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/cli/pipelines/{pipeline}/{phase}/{app}": {
      "delete": {
        "tags": [
          "CLI"
        ],
        "summary": "Delete an app",
        "description": "",
        "parameters": [
          {
            "name": "pipeline",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "phase",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "app",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": {
              "type": "http",
              "scheme": "bearer",
              "bearerFormat": "JWT"
            }
          }
        ]
      },
      "get": {
        "tags": [
          "CLI"
        ],
        "summary": "Get app details",
        "description": "",
        "parameters": [
          {
            "name": "pipeline",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "phase",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "app",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": {
              "type": "http",
              "scheme": "bearer",
              "bearerFormat": "JWT"
            }
          }
        ]
      }
    },
    "/cli/pipelines/{pipeline}/apps": {
      "get": {
        "tags": [
          "CLI"
        ],
        "summary": "Get all apps in a pipeline",
        "description": "",
        "parameters": [
          {
            "name": "pipeline",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": {
              "type": "http",
              "scheme": "bearer",
              "bearerFormat": "JWT"
            }
          }
        ]
      }
    },
    "/pipelines/{pipeline}/apps": {
      "get": {
        "tags": [
          "Pipeline"
        ],
        "summary": "Get all apps in a pipeline",
        "description": "",
        "parameters": [
          {
            "name": "pipeline",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/pipelines/{pipeline}/{phase}/{app}/restart": {
      "get": {
        "tags": [
          "Pipeline"
        ],
        "summary": "Restart an app",
        "description": "",
        "parameters": [
          {
            "name": "pipeline",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "phase",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "app",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}